# Optimize the parsing of special floats: infinity and nan.