Don't repeat the check on the first character in each string: 'i' or 'n'. Add a benchmark for special floats.
Supplement units test on special floats with SyntaxError cases.

Benchmarks on my M1 Mac show a -8% speedup:

$ benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/sugawarayuuta/refloat
                               │   old.txt   │              new.txt              │
                               │   sec/op    │   sec/op     vs base              │
ParseFloat64Special/refloat-10   8.895n ± 1%   8.197n ± 1%  -7.84% (p=0.002 n=6)